.github/scripts/architect/requirements.txt
google-generativeai
duckduckgo-search
requests
beautifulsoup4